+++ test/run-profile.sh
@@ -11,6 +11,7 @@ run_exe() {
     --cmd 'profile file *' \
     --cmd 'set verbosefile=verbose.log' \
     --cmd 'set verbose=9' \
+    --cmd "let g:clap_use_pure_python = $USE_PURE_PYTHON" \
     -c "call $test_fn()"
 }
 
@@ -72,6 +73,28 @@ run_bench() {
   bench_100000
 }
 
+run_all() {
+  run_once
+  run_multi
+  run_bench
+}
+
+test_python_and_rust() {
+  echo 'stats of pure Python fuzzy filter performance:' > stats.log
+  echo '' >> stats.log
+  USE_PURE_PYTHON=1
+  $1
+
+  echo '>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>' >> stats.log
+  echo '<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<' >> stats.log
+  echo '' >> stats.log
+
+  echo 'stats of Rust fuzzy filter performance:' >> stats.log
+  echo '' >> stats.log
+  USE_PURE_PYTHON=0
+  $1
+}
+
 help() {
   cat << EOF
 usage: $0 [OPTIONS]
@@ -88,19 +111,20 @@ if [ $# -eq 0 ]; then
   exit 1
 fi
 
-echo 'stats of fuzzy filter performance:' > stats.log
-echo '' >> stats.log
-
 for opt in "$@"; do
   case $opt in
     --help)  help      ;;
-    --once)  run_once  ;;
-    --multi) run_multi ;;
-    --bench) run_bench ;;
+    --once)
+      test_python_and_rust run_once
+      ;;
+    --multi)
+      test_python_and_rust run_multi
+      ;;
+    --bench)
+      test_python_and_rust run_bench
+      ;;
     --all)
-      run_once
-      run_multi
-      run_bench
+      test_python_and_rust run_all
       ;;
     *)
       echo "unknown option: $opt"